Introducing new branch management
We now have a hierarchy tree that contains all the ceph versions
including OS versions. Each OS version has its own `base` and `daemon`
directories. When it comes to Ubuntu, 14.04 remains the default where
16.04 has symlinks to 14.04. Only the base image version changes.

We have a script that can be executed from anywhere, even if you are
outside the repository it will work.
It will look for any new commits since the last tag and then will figure
out which ceph release was changed.

The script will do several things:

0. determine which ceph version and OS was modified
1. create a new branch name
2. delete the old branch
3. create a new branch
4. copy daemon and base files to the root
5. commit the new changes
6. tag the commit
7. push the new branch

The Docker Hub has been configured to build a new image for every tag we
create.
Since the approprieate commit (in the proper branch) is tagged
a build will automatically be triggered.

Signed-off-by: Sébastien Han <seb@redhat.com>
Co-Authored-By: Erwan Velu <erwan@redhat.com>